// Assets-Uncapped/front-portal-develop/src/services/api/organisation-users/models/MissingDocumentResponse.ts
export interface MissingDocumentResponse {
    /**
     * 
     * @type {string}
     * @memberof MissingDocumentResponse
     */
    documentType?: string;
    /**
     * 
     * @type {string}
     * @memberof MissingDocumentResponse
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof MissingDocumentResponse
     */
    subtitle?: string;
    /**
     * 
     * @type {string}
     * @memberof MissingDocumentResponse
     */
    description?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof MissingDocumentResponse
     */
    allowedFileFormats?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof MissingDocumentResponse
     */
    instructions?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof MissingDocumentResponse
     */
    exampleDocumentUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof MissingDocumentResponse
     */
    connectionType?: MissingDocumentResponseConnectionTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof MissingDocumentResponse
     */
    requestType?: MissingDocumentResponseRequestTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof MissingDocumentResponse
     */
    externalId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof MissingDocumentResponse
     */
    isOptional?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof MissingDocumentResponse
     */
    lastModificationDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof MissingDocumentResponse
     */
    documentRequestId?: string;
    /**
     * 
     * @type {AdditionalInfo}
     * @memberof MissingDocumentResponse
     */
    additionalInfo?: AdditionalInfo;
}


/**
 * @export
 */
export const MissingDocumentResponseConnectionTypeEnum = {
    Sales: 'SALES',
    Accounting: 'ACCOUNTING',
    Banking: 'BANKING'
} as const;
export type MissingDocumentResponseConnectionTypeEnum = typeof MissingDocumentResponseConnectionTypeEnum[keyof typeof MissingDocumentResponseConnectionTypeEnum];

/**
 * @export
 */
export const MissingDocumentResponseRequestTypeEnum = {
    Loan: 'LOAN',
    LineOfCredit: 'LINE_OF_CREDIT'
} as const;
export type MissingDocumentResponseRequestTypeEnum = typeof MissingDocumentResponseRequestTypeEnum[keyof typeof MissingDocumentResponseRequestTypeEnum];